v5.0.0 适配iPhoneX、XR、Xmax ；修复闪退问题；优化一些功能细节体验。

+++ CampusOfGLUt/Classes/Vendors/MWPhotoBrowser/MWPhotoBrowser.m
@@ -564,7 +564,7 @@
 }
 
 - (NSUInteger)supportedInterfaceOrientations {
-    return UIInterfaceOrientationMaskAll;
+    return (NSUInteger)UIInterfaceOrientationMaskAll;
 }
 
 - (void)willRotateToInterfaceOrientation:(UIInterfaceOrientation)toInterfaceOrientation duration:(NSTimeInterval)duration {
